@@ -17,13 +17,13 @@
  * @author volker.sorge@gmail.com (Volker Sorge)
  */
 
-import { Trie } from '../../speech-rule-engine/js/indexing/trie';
-import * as System from '../../speech-rule-engine/js/common/system';
-import { Variables } from '../../speech-rule-engine/js/common/variables';
-import { SpeechRuleEngine } from '../../speech-rule-engine/js/rule_engine/speech_rule_engine';
-import { SpeechRule } from '../../speech-rule-engine/js/rule_engine/speech_rule';
+import { Trie } from '../../speech-rule-engine/js/indexing/trie.js';
+import * as System from '../../speech-rule-engine/js/common/system.js';
+import { Variables } from '../../speech-rule-engine/js/common/variables.js';
+import { SpeechRuleEngine } from '../../speech-rule-engine/js/rule_engine/speech_rule_engine.js';
+import { SpeechRule } from '../../speech-rule-engine/js/rule_engine/speech_rule.js';
 
-import { JsonFile, TestPath, TestUtil } from '../base/test_util';
+import { JsonFile, TestPath, TestUtil } from '../base/test_util.js';
 
 namespace AnalyticsUtil {
   // Removes duplicates from a list in O(n).
