@@ -17,8 +17,8 @@
  */
 
 import * as process from 'process';
-import { AbstractTest } from '../classes/abstract_test';
-import { AbstractJsonTest } from '../classes/abstract_test';
+import { AbstractTest } from '../classes/abstract_test.js';
+import { AbstractJsonTest } from '../classes/abstract_test.js';
 
 enum Warning {
   NONE,
