@@ -18,10 +18,10 @@
  * @author Volker.Sorge@gmail.com (Volker Sorge)
  */
 
-import { annotators } from '../../speech-rule-engine/js/semantic_tree/semantic_annotations';
-import * as Semantic from '../../speech-rule-engine/js/semantic_tree/semantic';
+import { annotators } from '../../speech-rule-engine/js/semantic_tree/semantic_annotations.js';
+import * as Semantic from '../../speech-rule-engine/js/semantic_tree/semantic.js';
 
-import { AbstractJsonTest } from './abstract_test';
+import { AbstractJsonTest } from './abstract_test.js';
 
 export class ClearspeakAnnotationTest extends AbstractJsonTest {
   /**
