@@ -16,10 +16,10 @@
  * @author sorge@google.com (Volker Sorge)
  */
 
-import * as DomUtil from '../../speech-rule-engine/js/common/dom_util';
-import * as XpathUtil from '../../speech-rule-engine/js/common/xpath_util';
+import * as DomUtil from '../../speech-rule-engine/js/common/dom_util.js';
+import * as XpathUtil from '../../speech-rule-engine/js/common/xpath_util.js';
 
-import { AbstractJsonTest } from '../classes/abstract_test';
+import { AbstractJsonTest } from '../classes/abstract_test.js';
 
 export class HtmlTest extends AbstractJsonTest {
   /**
