@@ -20,17 +20,17 @@
 import * as fs from 'fs';
 import * as path from 'path';
 
-import * as DomUtil from '../../speech-rule-engine/js/common/dom_util';
-import * as Enrich from '../../speech-rule-engine/js/enrich_mathml/enrich';
-import * as Semantic from '../../speech-rule-engine/js/semantic_tree/semantic';
-import { SemanticNode } from '../../speech-rule-engine/js/semantic_tree/semantic_node';
-import { SemanticTree } from '../../speech-rule-engine/js/semantic_tree/semantic_tree';
+import * as DomUtil from '../../speech-rule-engine/js/common/dom_util.js';
+import * as Enrich from '../../speech-rule-engine/js/enrich_mathml/enrich.js';
+import * as Semantic from '../../speech-rule-engine/js/semantic_tree/semantic.js';
+import { SemanticNode } from '../../speech-rule-engine/js/semantic_tree/semantic_node.js';
+import { SemanticTree } from '../../speech-rule-engine/js/semantic_tree/semantic_tree.js';
 
-import * as tu from '../base/test_util';
-import * as TestFactory from '../classes/test_factory';
-import { addActual } from './fill_tests';
-import { Tex2Mml } from './tex_transformer';
-import { AbstractTransformer, Transformer } from './transformers';
+import * as tu from '../base/test_util.js';
+import * as TestFactory from '../classes/test_factory.js';
+import { addActual } from './fill_tests.js';
+import { Tex2Mml } from './tex_transformer.js';
+import { AbstractTransformer, Transformer } from './transformers.js';
 
 /* ********************************************************* */
 /*
